bug : 사용자 로그인 기능, 유저 리스트 보는 기능이 동작하지 않는 버그 수정
