+++ src/main/java/http/Cookies.java
@@ -18,6 +18,7 @@ public class Cookies {
     }
 
     public Cookies(String values) {
+        cookies = new HashMap<>();
         for (String cookie : values.split(COOKIE_TOKENIZER)) {
             String[] c = cookie.split(COOKIE_NAME_VALUE_TOKENIZER);
             if (c.length != 2) {
